+++ Packages/com.radtad2.sceneservice/Runtime/SceneGroup.cs
@@ -22,11 +22,18 @@ namespace SceneService
         /// All scenes that the active scene depends on.
         /// </summary>
         [SerializeField] private List<SceneReference> dependencies;
+
+        [SerializeField] private List<string> tags;
         
         /// <summary>
         /// A readonly list of all dependencies.
         /// </summary>
         public IReadOnlyList<SceneReference> Dependencies => dependencies;
+        
+        /// <summary>
+        /// A readonly list of all tags.
+        /// </summary>
+        public IReadOnlyList<string> Tags => tags;
         
         /// <summary>
         /// Gets a new list of all scenes in this group.